made building exe optional, renamed rlVector/rlPoint

+++ build.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+build_executable="OFF"
 build_lib="OFF"
 build_tests="OFF"
 build_all=false
@@ -8,10 +9,13 @@ build_debug=false
 build_sanitized=false
 defaults=true
 
-while getopts 'altrds' OPTION; do
+while getopts 'aeltrds' OPTION; do
   case "$OPTION" in
     a)
       build_all=true
+      ;;
+    e)
+      build_executable="ON"
       ;;
     l)
       build_lib="ON"
@@ -39,13 +43,22 @@ while getopts 'altrds' OPTION; do
   esac
 done
 
+if [ "$OPTIND" = 1 ] ; then
+
+  echo "No options given - I don't know what to do!"
+  echo "Please tell me what to build..."
+
+  exit 1
+
+fi
+
 rm -rf build
 
 if [ "$build_all" = true ] ; then
 
-  cmake -B build/release -DCMAKE_BUILD_TYPE=RELEASE -DBUILD_LIB=ON -DBUILD_TESTS=ON
-  cmake -B build/debug -DCMAKE_BUILD_TYPE=DEBUG -DBUILD_LIB=ON -DBUILD_TESTS=ON
-  cmake -B build/sanitized -DCMAKE_BUILD_TYPE=SANITIZED -DBUILD_LIB=ON -DBUILD_TESTS=ON
+  cmake -B build/release -DCMAKE_BUILD_TYPE=RELEASE -DBUILD_EXECUTABLE=ON -DBUILD_LIB=ON -DBUILD_TESTS=ON
+  cmake -B build/debug -DCMAKE_BUILD_TYPE=DEBUG -DBUILD_EXECUTABLE=ON -DBUILD_LIB=ON -DBUILD_TESTS=ON
+  cmake -B build/sanitized -DCMAKE_BUILD_TYPE=SANITIZED -DBUILD_EXECUTABLE=ON -DBUILD_LIB=ON -DBUILD_TESTS=ON
 
   cmake --build build/release
   cmake --build build/debug
@@ -57,7 +70,7 @@ fi
 
 if [ "$build_release" = true ] ; then
 
-  cmake -B build/release -DCMAKE_BUILD_TYPE=RELEASE -DBUILD_LIB="$build_lib" -DBUILD_TESTS="$build_tests"
+  cmake -B build/release -DCMAKE_BUILD_TYPE=RELEASE -DBUILD_EXECUTABLE="$build_executable" -DBUILD_LIB="$build_lib" -DBUILD_TESTS="$build_tests"
 
   cmake --build build/release
 
@@ -65,7 +78,7 @@ fi
 
 if [ "$build_debug" = true ] ; then
 
-  cmake -B build/debug -DCMAKE_BUILD_TYPE=DEBUG -DBUILD_LIB="$build_lib" -DBUILD_TESTS="$build_tests"
+  cmake -B build/debug -DCMAKE_BUILD_TYPE=DEBUG -DBUILD_EXECUTABLE="$build_executable" -DBUILD_LIB="$build_lib" -DBUILD_TESTS="$build_tests"
 
   cmake --build build/debug
 
@@ -73,7 +86,7 @@ fi
 
 if [ "$build_sanitized" = true ] ; then
 
-  cmake -B build/sanitized -DCMAKE_BUILD_TYPE=SANITIZED -DBUILD_LIB="$build_lib" -DBUILD_TESTS="$build_tests"
+  cmake -B build/sanitized -DCMAKE_BUILD_TYPE=SANITIZED -DBUILD_EXECUTABLE="$build_executable" -DBUILD_LIB="$build_lib" -DBUILD_TESTS="$build_tests"
 
   cmake --build build/sanitized
 
@@ -81,7 +94,7 @@ fi
 
 if [ "$defaults" = true ] ; then
 
-  cmake -B build/release -DCMAKE_BUILD_TYPE=RELEASE -DBUILD_LIB="$build_lib" -DBUILD_TESTS="$build_tests"
+  cmake -B build/release -DCMAKE_BUILD_TYPE=RELEASE -DBUILD_EXECUTABLE="$build_executable" -DBUILD_LIB="$build_lib" -DBUILD_TESTS="$build_tests"
 
   cmake --build build/release
